requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
httpx>=0.27.0
playwright>=1.44.0
playwright-stealth>=2.0.0
python-dotenv>=1.0.0
